Extract agent configuration types to domain model
